Conflicts:
	MonoGame.Framework/MonoGame.Framework.Windows8.csproj
	MonoGame.Framework/Windows/GamerServices/Guide.cs

Merged with new Protobuild process.

+++ MonoGame.Framework/Graphics/DisplayMode.cs
@@ -49,7 +49,7 @@ namespace Microsoft.Xna.Framework.Graphics
         #region Properties
         
         public float AspectRatio {
-            get { return width / height; }
+            get { return (float)width / (float)height; }
         }
 
         public SurfaceFormat Format {
